requirements.txt
Flask>=2.2
matplotlib>=3.6
google-api-python-client>=2.70
google-auth>=2.20
google-auth-httplib2>=0.1
python-dotenv>=1.0